backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
httpx
passlib[bcrypt]
bcrypt==4.0.1
pyjwt
alembic
psycopg2-binary
asyncpg
email-validator
python-multipart
pytz
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
cryptography
